+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## 0.0.1
-
-* TODO: Describe initial release.
+- Create new package from https://github.com/PavelS0/docx_template_dart
+## 0.0.2
+- Add test
+- Fixed null safety for xml package
+- Added function get all tags